+++ examples/data-sources/ceph_cluster_user/data-source.tf
@@ -1,0 +1,7 @@
+data "ceph_cluster_user" "admin" {
+  entity = "client.admin"
+}
+
+output "admin_capabilities" {
+  value = data.ceph_cluster_user.admin.capabilities
+}